% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papermill.R
\name{pm_display}
\alias{pm_display}
\title{Papermill Display}
\usage{
pm_display(name, obj)
}
\arguments{
\item{name}{an alias for the obj}

\item{obj}{any object which can be serialized using repr}
}
\description{
Capture the display of an R command.
}
\details{
Obj is any object that can be serialized using
\href{https://cran.r-project.org/web/packages/repr/index.html}{repr},
for example text, images, and data.frames.
If obj is a plot from R's base graphics you must call \code{renderPlot} on it.
If obj is a plot from ggplot you must print it then call \code{renderPlot}.
}
\examples{

\dontrun{
plot(1:10)
p = recordPlot()
pm_display("myplot", p)
}

\dontrun{
require(ggplot2)
df = data.frame(x = 1:5, y = 1:5)
myplot = ggplot(df,
aes(x = x, y = y)) +
geom_point()
pm_display("myplot", recordPlot(print(myplot)))
}
}
