% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papermill.R
\name{pm_record}
\alias{pm_record}
\title{Record data into Papermill}
\usage{
pm_record(keys, values)
}
\arguments{
\item{keys}{a vector of character type. This is used
to name the objects stored in value}

\item{value}{a list of values to record.
Limited to characters, logicals, numerics.
The list does not have to have homogenous data types.}
}
\description{
Publishes a mimebundle to Papermills' metadata.
}
\details{
The keys and values that are recorded can be accessed again
in another notebook, for example for the purpose of
meta analyzing multiple notebooks
}
\examples{

\dontrun{
keys = letters[1:5]
values = as.list(1:5)
pm_record(keys, values)
}
}
